backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
groq
pydantic>=2.0
